Refactor RealmCollection and eliminate lots of duplicated code
Define everything which can be as extensions on RealmCollection rather than in
each of the collection implementations.

Add RealmCollectionImpl which defines implementations of all of the things
which forward to the backing RLMCollection, and adjust all of the Swift
collections to be backed by a RLMCollection.

Add some missing things to RLMCollection to make this possible.

Change AnyRealmCollection to wrap a RLMCollection instead of a RealmCollection.
This should not have any user-visible effect other than improved performance
and a slightly different result from the description method.