Adjust where bridging is done for query values
Use RLMStatelessAccessor::unbox() when reading values for queries rather than a
more limited query-specific version of that. This both eliminates some
duplicated code and makes the query builder support types which require
bridging, such as PersistableEnums (and custom type mappings) in collections.

Replace the explicit handling for PersistableEnum in Query with moving the
constraint checks to PersistedType where applicable. Some operations can't
really be performed on the enum type (e.g. sum and average) but might still be
meaningful queries to write, so add a `.rawValue` modifier to `Query` to
support querying the raw value instead of the enum. This has no effect on the
constructed NSPredicate as the above change makes them interchangable at that
layer.

Adding tests for TSQs on enum types required significant refactoring to
QueryTests.swift.gyb: first to make it possible to construct the tests at all,
and then to reduce the size of the generated file. After adding the new types
to the existing tests QueryTests.swift was over 50k lines of code, took
multiple minutes to compile, and locked up Xcode if you made the mistake of
opening the file. Pushing a lot of the looping from GYB to Swift makes the
tests a bit worse, but the status quo was unworkable.